search bar works. Need to select specific data to show instead of full JSON data set
