+++ views/index.ejs
@@ -8,7 +8,7 @@
 <body>
     <h1>Zacks Ballin' AF Weather Data Spewin App</h1>
     <h4>Enter your zipcode below to get REAL up to date weather conditions for where you live</h4>
-    <form>
+    <form action = "/weather" method="get">
         <input type="search" id="searchZip" name="searchBar" placeholder="Zip Code">
         <button for="searchBar">Search</button>
     </form>
